Added classes, tests, Spring Security, authorization, authentication

+++ src/main/java/org/example/services/impl/OrderItemServiceImpl.java
@@ -1,90 +1,112 @@
 package org.example.services.impl;
 
 import jakarta.transaction.Transactional;
+import org.example.dto.OrderItemCreateDto;
 import org.example.dto.OrderItemDto;
-import org.example.exception.OrderItemNotFoundException;
-import org.example.exception.OrderNotFoundException;
-import org.example.exception.ProductNotFoundException;
-import org.example.exception.errorMessage.ErrorMessage;
+import org.example.enums.Role;
+import org.example.exceptions.OrderItemNotFoundException;
+import org.example.exceptions.OrderNotFoundException;
+import org.example.exceptions.ProductNotFoundException;
+import org.example.exceptions.errorMessage.ErrorMessage;
 import org.example.mappers.OrderItemMapper;
 import org.example.models.Order;
 import org.example.models.OrderItem;
 import org.example.models.Product;
+import org.example.models.User;
 import org.example.repositories.OrderItemRepository;
 import org.example.repositories.OrderRepository;
 import org.example.repositories.ProductRepository;
+import org.example.security.SecurityUtils;
 import org.example.services.interfaces.OrderItemService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 public class OrderItemServiceImpl implements OrderItemService {
     private final OrderItemRepository orderItemRepository;
     private final OrderRepository orderRepository;
     private final ProductRepository productRepository;
     private final OrderItemMapper orderItemMapper;
+    private final SecurityUtils securityUtils;
 
-    @Autowired
     public OrderItemServiceImpl(OrderItemRepository orderItemRepository,
                                 OrderRepository orderRepository,
                                 ProductRepository productRepository,
-                                OrderItemMapper orderItemMapper) {  // Внедрение маппера
+                                OrderItemMapper orderItemMapper,
+                                SecurityUtils securityUtils) {
         this.orderItemRepository = orderItemRepository;
         this.orderRepository = orderRepository;
         this.productRepository = productRepository;
-        this.orderItemMapper = orderItemMapper;  // Инициализация
+        this.orderItemMapper = orderItemMapper;
+        this.securityUtils = securityUtils;
     }
 
+    private User getCurrentUser() {
+        return securityUtils.getCurrentUser();
+    }
     @Transactional
     @Override
-    public OrderItem createOrderItem(OrderItemDto orderItemDto, Long orderId) {
-        // Получаем заказ по ID
+    public OrderItemDto createOrderItem(OrderItemCreateDto orderItemCreateDto, Long orderId) {
+        User currentUser = getCurrentUser();
+
+        Order order = orderRepository.findById(orderId)
+                .orElseThrow(() -> new OrderNotFoundException(ErrorMessage.ORDER_NOT_FOUND));
+        checkOrderOwnership(order, currentUser);
+
+
+        Product product = getProductById(orderItemCreateDto.getProductId());
+
+
+        OrderItem orderItem = orderItemMapper.toEntity(orderItemCreateDto, product);
+        orderItem.setOrder(order);
+
+
+        orderItem = orderItemRepository.save(orderItem);
+
+
+        updateOrderTotalAmount(order);
+
+        return orderItemMapper.toDto(orderItem);
+    }
+
+    @Override
+    public List<OrderItemDto> getOrderItemsByOrderId(Long orderId) {
+        User currentUser = getCurrentUser();
+
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new OrderNotFoundException(ErrorMessage.ORDER_NOT_FOUND));
 
-        // Получаем продукт по ID
-        Product product = productRepository.findById(orderItemDto.getProductId())
-                .orElseThrow(() -> new ProductNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND));
-
-        // Создаем новый OrderItem из DTO и Product
-        OrderItem orderItem = orderItemMapper.toEntity(orderItemDto, product);
-        orderItem.setOrder(order);
-
-        // Сохраняем новый OrderItem
-        orderItem = orderItemRepository.save(orderItem);
-
-        // Обновляем общую сумму заказа
-        updateOrderTotalAmount(order);
-
-        return orderItem;
-    }
-
-    @Override
-    public List<OrderItem> getOrderItemsByOrderId(Long orderId) {
-        return orderItemRepository.findByOrderId(orderId);
+        checkOrderOwnership(order, currentUser);
+        return orderItemRepository.findByOrderId(orderId).stream()
+                .map(orderItemMapper::toDto)
+                .collect(Collectors.toList());
     }
     @Transactional
     @Override
-    public OrderItem updateOrderItem(Long orderItemId, OrderItemDto orderItemDto) {
+    public OrderItemDto updateOrderItem(Long orderItemId, OrderItemCreateDto orderItemCreateDto) {
+        User currentUser = getCurrentUser();
         OrderItem orderItem = orderItemRepository.findById(orderItemId)
                 .orElseThrow(() -> new OrderItemNotFoundException(ErrorMessage.ORDER_ITEM_NOT_FOUND));
+        Order order = orderItem.getOrder();
 
-        // Получаем продукт по ID
-        Product product = productRepository.findById(orderItemDto.getProductId())
-                .orElseThrow(() -> new ProductNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND));
+        checkOrderOwnership(order, currentUser);
 
-        // Обновляем данные OrderItem
-        orderItemMapper.updateEntityFromDto(orderItemDto, orderItem, product);
 
-        // Сохраняем обновленный OrderItem
+        Product product = getProductById(orderItemCreateDto.getProductId());
+
+
+        orderItemMapper.updateEntityFromCreateDto(orderItemCreateDto, orderItem, product);
+
+
         orderItem = orderItemRepository.save(orderItem);
 
-        // Обновляем общую сумму заказа
+
         updateOrderTotalAmount(orderItem.getOrder());
 
-        return orderItem;
+        return orderItemMapper.toDto(orderItem);
     }
     private void updateOrderTotalAmount(Order order) {
         BigDecimal totalAmount = order.getOrderItems().stream()
@@ -98,13 +120,30 @@ public class OrderItemServiceImpl implements OrderItemService {
     @Transactional
     @Override
     public void deleteOrderItem(Long orderItemId) {
+        User currentUser = getCurrentUser();
         OrderItem orderItem = orderItemRepository.findById(orderItemId)
                 .orElseThrow(() -> new OrderItemNotFoundException(ErrorMessage.ORDER_ITEM_NOT_FOUND));
 
         Order order = orderItem.getOrder();
-        orderItemRepository.delete(orderItem);
 
-        // Обновляем общую сумму заказа
+        checkOrderOwnership(order, currentUser);
+
+        order.getOrderItems().remove(orderItem);
+
         updateOrderTotalAmount(order);
+
+        orderItemRepository.delete(orderItem);
+    }
+
+
+
+    private Product getProductById(Long productId) {
+        return productRepository.findById(productId)
+                .orElseThrow(() -> new ProductNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND));
+    }
+    private void checkOrderOwnership(Order order, User user) {
+        if (!order.getUser().getId().equals(user.getId()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new org.springframework.security.access.AccessDeniedException(ErrorMessage.ACCESS_DENIED);
+        }
     }
 }
